--- Bug Report and Fix.txt
@@ -1,18 +1,0 @@
-Append bug report and fixes to the end of this file. Please use the following format and separate each bug report with dashed lines:
-
-Example bug report:
-----------------------------------------------------
-Bug report
-ID #: 1  (just pick an unused number)
-Date: 4/26/2017
-Bug description: Player enters "n" and gets message "I beg your pardon...", but it seems that the game should recognize this command.
-
-----------------------------------------------------
-
-
-Example bug fix:
-----------------------------------------------------
-Bug fix
-ID #: 1
-Date: 4/27/2017
-----------------------------------------------------
